tests for GetKeyspace in SessionTests.cs
Added "Session_Get_Keyspace" to SessionTests.cs
Moved tests "ChangeKeyspace_SetsKeyspace" and "Cluster_ConnectAsync_Should_Create_A_Session_With_Keyspace_Set" from BrokenIntegrationTests/Core/SessionTests.cs
Also modified "Session_Keyspace_Empty_On_Connect" and "Session_Use_Statement_Changes_Keyspace_Case_Insensitive" to assert localSession.Keyspace is correct
