+++ src/Cassandra.IntegrationTests/Core/SessionTests.cs
@@ -15,7 +15,8 @@
 //
 
 using System.Diagnostics;
-
+using System.Threading.Tasks;
+using Cassandra.IntegrationTests.TestBase;
 using Cassandra.Tests;
 
 using NUnit.Framework;
@@ -46,6 +47,7 @@ namespace Cassandra.IntegrationTests.Core
             {
                 var localSession = localCluster.Connect("");
                 localSession.Execute("SELECT * FROM system.local WHERE key='local'");
+                Assert.That(localSession.Keyspace, Is.Null);
             });
         }
 
@@ -88,6 +90,7 @@ namespace Cassandra.IntegrationTests.Core
                     localSession.Execute("select * from local");
                 }
             });
+            Assert.That(localSession.Keyspace, Is.EqualTo("system"));
         }
 
         [Test]
@@ -120,6 +123,92 @@ namespace Cassandra.IntegrationTests.Core
             var localSession = localCluster.Connect();
             var ex = Assert.Throws<InvalidQueryException>(() => localSession.ChangeKeyspace("THIS_KEYSPACE_DOES_NOT_EXIST_EITHER"));
             Assert.True(ex.Message.ToLower().Contains("keyspace"));
+        }
+
+        [Test]
+        public void ChangeKeyspace_SetsKeyspace()
+        {
+            var localCluster = GetNewTemporaryCluster();
+            var localSession = localCluster.Connect();
+            localSession.CreateKeyspace(KeyspaceName, null, false);
+            localSession = localCluster.Connect();
+            Assert.IsNull(localSession.Keyspace);
+            localSession.ChangeKeyspace(KeyspaceName);
+            Assert.IsNotNull(localSession.Keyspace);
+            Assert.AreEqual(KeyspaceName, localSession.Keyspace);
+        }
+
+        [Test]
+        public async Task Cluster_ConnectAsync_Should_Create_A_Session_With_Keyspace_Set()
+        {
+            const string query = "SELECT * FROM local";
+            // Using a default keyspace
+            using (var cluster = ClusterBuilder().AddContactPoint(TestCluster.InitialContactPoint)
+                                           .WithDefaultKeyspace("system").Build())
+            {
+                ISession session = await cluster.ConnectAsync().ConfigureAwait(false);
+                Assert.DoesNotThrowAsync(async () =>
+                    await session.ExecuteAsync(new SimpleStatement(query)).ConfigureAwait(false));
+                Assert.DoesNotThrowAsync(async () =>
+                    await session.ExecuteAsync(new SimpleStatement(query)).ConfigureAwait(false));
+                await cluster.ShutdownAsync().ConfigureAwait(false);
+            }
+
+            // Setting the keyspace on ConnectAsync
+            using (var cluster = ClusterBuilder().AddContactPoint(TestCluster.InitialContactPoint).Build())
+            {
+                ISession session = await cluster.ConnectAsync("system").ConfigureAwait(false);
+                Assert.DoesNotThrowAsync(async () =>
+                    await session.ExecuteAsync(new SimpleStatement(query)).ConfigureAwait(false));
+                Assert.DoesNotThrowAsync(async () =>
+                    await session.ExecuteAsync(new SimpleStatement(query)).ConfigureAwait(false));
+                await cluster.ShutdownAsync().ConfigureAwait(false);
+            }
+
+            // Without setting the keyspace
+            using (var cluster = ClusterBuilder().AddContactPoint(TestCluster.InitialContactPoint).Build())
+            {
+                ISession session = await cluster.ConnectAsync().ConfigureAwait(false);
+                Assert.DoesNotThrowAsync(async () =>
+                    await session.ExecuteAsync(new SimpleStatement("SELECT * FROM system.local WHERE key='local'"))
+                                 .ConfigureAwait(false));
+                await cluster.ShutdownAsync().ConfigureAwait(false);
+            }
+        }
+
+        [Test]
+        public void Session_Get_Keyspace() {
+            var localCluster = GetNewTemporaryCluster();
+
+            using (var localSession = localCluster.Connect())
+            {
+                Assert.That(localSession.Keyspace, Is.Null);
+            }
+
+            using (var localSession = localCluster.Connect(""))
+            {
+                Assert.That(localSession.Keyspace, Is.Null);
+            }
+
+            using (var localSession = localCluster.Connect("system"))
+            {
+                Assert.That(localSession.Keyspace, Is.EqualTo("system"));
+            }
+
+            using (var localSession = localCluster.Connect())
+            {
+                Assert.That(localSession.Keyspace, Is.Null);
+                localSession.ChangeKeyspace("system");
+                Assert.That(localSession.Keyspace, Is.EqualTo("system"));
+            }
+
+            using (var localSession = localCluster.Connect())
+            {
+                Assert.That(localSession.Keyspace, Is.Null);
+                localSession.Execute("USE SyStEm");
+                Assert.That(localSession.Keyspace, Is.EqualTo("system"));
+                Assert.AreNotEqual(localSession.Keyspace, "SyStEm");
+            }
         }
     }
 }
